+++ 221701435/README.md
@@ -1,4 +1,25 @@
-# InfectStatistic-学号
-疫情统计
+# InfectStatistic-221701435
+### 运行
 
-描述你的项目，包括如何运行、功能简介、作业链接、博客链接等
+在命令行输入指令 java infectStatic list -log ../log -out ../result/output.txt
+
+//可附加参数
+
+```
+-type sp ip dead cure  //限定只输出疑似、感染、死亡、治愈数据。
+-province 西藏 香港  //限定只输出这些省份。若存在限定省份，则默认不输出“全国”。
+-date 2020-02-05    //只统计文件至该日期。
+```
+
+### 功能简介
+
+统计每天的日志文件，并汇总统计结果
+
+### 作业链接
+
+
+
+### 博客链接
+
+
+
