+++ 221701435/result/ListOut1.txt
@@ -1,5 +1,3 @@
-å…¨å›½ æ„ŸæŸ“æ‚£è€…15äºº ç–‘ä¼¼æ‚£è€…22äºº æ²»æ„ˆ2äºº æ­»äº¡1äºº
-ç¦å»º æ„ŸæŸ“æ‚£è€…5äºº ç–‘ä¼¼æ‚£è€…7äºº æ²»æ„ˆ0äºº æ­»äº¡0äºº
-æ¹–åŒ— æ„ŸæŸ“æ‚£è€…10äºº ç–‘ä¼¼æ‚£è€…15äºº æ²»æ„ˆ2äºº æ­»äº¡1äºº
-// è¯¥æ–‡æ¡£å¹¶éžçœŸå®žæ•°æ®ï¼Œä»…ä¾›æµ‹è¯•ä½¿ç”¨
-// å‘½ä»¤ï¼šlist -log D:\log\ -out D:\ListOut1.txt -date 2020-01-22
+È«¹ú ¸ÐÈ¾»¼Õß15ÈË ÒÉËÆ»¼Õß22ÈË ÖÎÓú2ÈË ËÀÍö1ÈË
+¸£½¨ ¸ÐÈ¾»¼Õß5ÈË ÒÉËÆ»¼Õß7ÈË ÖÎÓú0ÈË ËÀÍö0ÈË
+ºþ±± ¸ÐÈ¾»¼Õß10ÈË ÒÉËÆ»¼Õß15ÈË ÖÎÓú2ÈË ËÀÍö1ÈË
